fix download sorting algorithm in DLStrategyA

+++ src/main/scala/it/polimi/genomics/metadata/downloader_transformer/one_k_genomes/DLStrategyA.scala
@@ -123,9 +123,7 @@ class DLStrategyA extends Downloader {
   private def fetchUpdatesForVariants(treeLocalPath: String, dataset: Dataset): Unit = {
     //        3
     val variantRecords_asLines = DatasetInfo.latestVariantsRecords(treeLocalPath, dataset)
-    val variantRecords = parseVariantRecords(variantRecords_asLines).sortBy(record => record._2)  // sorting by size - from smallest to biggest - is done for debug purposes
-    //TODO print list of variant records sorted by size to check sorting... Maybe sorting on strings works differently
-    // adding .toLong in sortBy could do the trick
+    val variantRecords = parseVariantRecords(variantRecords_asLines).sortBy(record => record._2.toLong)  // sorting by size - from smallest to biggest - is done for debug purposes
     if (variantRecords.isEmpty) {
       throw new IllegalStateException("DOWNLOAD CAN'T CONTINUE: UNABLE TO EXTRACT VARIANT SET FROM TREE LOCAL FILE")
     } else {
